requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
mlflow
fastapi
uvicorn[standard]
httpx
beautifulsoup4
lxml
pydantic
python-dotenv
rich
orjson
